agent/run-context: Set unit name on systemd scopes
Name them `ptyxis-spawn-<RandomUUID>.scope`, following similar naming in
VTE and tmux.

This identifies them as Ptyxis' scopes and allows applying configuration
to them with snippets in `systemd/user/ptyxis-spawn-.scope.d`.

In the future we could change the prefix to `ptyxis-spawn-<ProfileID>-`
to allow separate configuration for each profile.

Fixes: https://gitlab.gnome.org/chergert/ptyxis/-/issues/382