style: use padding instead of margin
The original change here was in 90e72ca7dd8d64650c2dc3002e8a199429ed0cf6
to help with performance. However, that seems (at least everywhere I tested
today) to be fine with NGL now.

We can go back to using padding because that improves the ability to select
text near the edge of the window border.

Fixes: #381