terminal: avoid Document Portal paths during Drag'n'Drop
While Ptyxis is running in a Flatpak container, the terminal we are pasting
into likely is not. Therefore, we don't want the URI that has been shared
with us inside of Flatpak but what it would be on the outside.

This tries to resolve that to the outside path in the same manner we did
in Text Editor and elsewhere.

Related: #479